requirements.txt
# ============================================
# Core Framework
# ============================================
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1

# ============================================
# Database
# ============================================
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1

# ============================================
# Authentication & Security
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator==2.1.1
cryptography==42.0.5

# ============================================
# File Processing
# ============================================
python-multipart==0.0.9
pandas==2.2.1
openpyxl==3.1.2
numpy==1.26.4

# ============================================
# PDF Generation
# ============================================
fpdf==1.7.2
Pillow==10.3.0

# ============================================
# WebSockets
# ============================================
websockets==12.0

# ============================================
# Production Server
# ============================================
gunicorn==21.2.0